requirements.txt
pytest==7.3.1
imageai==3.0.3
matplotlib==3.5.1
mpi4py==3.1.4
mpi4py_mpich==3.1.2
MySQL-python==1.2.5
numpy==1.21.5
object_detection==0.0.3
object_detection_0.1==0.1
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
pycocotools==2.0.6
pylabel==0.1.48
scipy==1.8.0
segment_anything==1.0
skimage==0.0
stochopy==2.2.0
tensorflow==2.12.0
